refactor(config): update import paths to use hash-based aliases for consistency
